+++ public/userscripts/manga-tracker.user.js
@@ -23,8 +23,9 @@
 // @include      /^http:\/\/mngcow\.co\/[a-zA-Z0-9_]+\/[0-9]+\/([0-9]+\/)?$/
 // @include      /^https:\/\/jaiminisbox\.com\/reader\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^https:\/\/kobato\.hologfx\.com\/reader\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
-// @updated      2017-01-14
-// @version      1.3.3
+// @include      /^http:\/\/www\.demonicscans\.com\/FoOlSlide\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
+// @updated      2017-01-15
+// @version      1.3.4
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -1200,6 +1201,38 @@ let sites = {
 		}
 	}),
 
+	'jaiminisbox.com' : extendSite({
+		setObjVars : function() {
+			this.title       = this.segments[3];
+			this.chapter     = this.segments[4] + '/' + this.segments[5] + '/' + this.segments[6] + (this.segments[7] && this.segments[7] !== 'page' ? '/' + this.segments[7] : '');
+
+			this.title_url   = this.https+'://jaiminisbox.com/reader/series/'+this.title;
+			this.chapter_url = this.https+'://jaiminisbox.com/reader/read/'+this.title+'/'+this.chapter;
+
+			this.chapterList        = generateChapterList($('.topbar_left > .tbtitle:eq(2) > ul > li > a').reverseObj(), 'href');
+			//FIXME: The chapterList isn't properly ordered for series that have chapters in and outside volumes. - https://reader.seaotterscans.com/series/sss/
+			console.log(this.chapter_url+'/');
+			this.chapterListCurrent = this.chapter_url+'/';
+
+			// this.viewerChapterName     = $('.selectChapter:first > option:selected').text().trim();
+			this.viewerTitle           = $('.topbar_left > .dropdown_parent > .text a').text();
+			this.viewerCustomImageList = $('#content').find('> script:first').html().match(/(https?:\\\/\\\/[^"]+)/g).filter(function(value, index, self) {
+				return self.indexOf(value) === index;
+			}).map(function(e) {
+				return e.replace(/\\/g, '');
+			});
+			this.page_count = this.viewerCustomImageList.length;
+		   },
+		   postSetupTopBar : function() {
+			$('.topbar_left > .tbtitle:eq(2)').remove();
+			$('.topbar_right').remove();
+		},
+		preSetupViewer : function(callback) {
+			$('#page').replaceWith($('<div/>', {id: 'viewer'})); //Set base viewer div
+			callback(true, true);
+		}
+	}),
+
 	//Doki Fansubs
 	'kobato.hologfx.com' : extendSite({
 		setObjVars : function() {
@@ -1232,17 +1265,16 @@ let sites = {
 		}
 	}),
 
-	'jaiminisbox.com' : extendSite({
+	'www.demonicscans.com' : extendSite({
 		setObjVars : function() {
 			this.title       = this.segments[3];
 			this.chapter     = this.segments[4] + '/' + this.segments[5] + '/' + this.segments[6] + (this.segments[7] && this.segments[7] !== 'page' ? '/' + this.segments[7] : '');
 
-			this.title_url   = this.https+'://jaiminisbox.com/reader/series/'+this.title;
-			this.chapter_url = this.https+'://jaiminisbox.com/reader/read/'+this.title+'/'+this.chapter;
+			this.title_url   = this.https+'://www.demonicscans.com/FoOlSlide/series/'+this.title;
+			this.chapter_url = this.https+'://www.demonicscans.com/FoOlSlide/read/'+this.title+'/'+this.chapter;
 
 			this.chapterList        = generateChapterList($('.topbar_left > .tbtitle:eq(2) > ul > li > a').reverseObj(), 'href');
 			//FIXME: The chapterList isn't properly ordered for series that have chapters in and outside volumes. - https://reader.seaotterscans.com/series/sss/
-			console.log(this.chapter_url+'/');
 			this.chapterListCurrent = this.chapter_url+'/';
 
 			// this.viewerChapterName     = $('.selectChapter:first > option:selected').text().trim();